# Remove unnecessary cart present and improve performance