--- a/controllers/front/CartController.php
+++ b/controllers/front/CartController.php
@@ -572,9 +572,8 @@ class CartControllerCore extends FrontController
     public function getTemplateVarPage()
     {
         $page = parent::getTemplateVarPage();
-        $presented_cart = $this->cart_presenter->present($this->context->cart);
 
-        if (count($presented_cart['products']) == 0) {
+        if (!$this->context->cart->hasProducts()) {
             $page['body_classes']['cart-empty'] = true;
         }