| Questions         | Answers
| ----------------- | -------------------------------------------------------
| Branch?           | 8.2.x
| Description?      | Boosts performance a lot on cart page by avoiding useless presenting of cart just to add one body class.
| Type?             | refacto
| Category?         | FO
| BC breaks?        | no
| Deprecations?     | no
| How to test?      | See below. Live shop with this change - https://eshop-franke.cz/
| UI Tests          | 
| Fixed issue or discussion?     | 
| Related PRs       | 
| Sponsor company   | 

### How to test
- Open https://www.domain.com/cart?action=show
- With products in cart, verify that there is no `cart-empty` class 
`<body id="cart" class="lang-cs country-cz currency-czk layout-full-width page-cart tax-display-enabled">`
- With empty cart, verify that there is `cart-empty` class 
`<body id="cart" class="lang-cs country-cz currency-czk layout-full-width page-cart tax-display-enabled cart-empty">`


### Performance gain with empty cart on 8.1.7
62 ms -> 54 ms
136 queries -> 121 queries

### Performance gain with 5 items in cart on 8.1.7
240 ms -> 216 ms
645 queries -> 569 queries