@Hlavtox 2024-11-19
@florine2623 @kpodemski Auto tests somehow don't work at all on this branch and 43/43 end up with timeout.

I think it's a safe change easy to test manually. Maybe @kpodemski you could test it quick? It's just that one HTML class.